Refactored so that we will only create one root logger
...and therefore, only initialize Google Cloud Logging's logging-winston once
