+++ src/gateway/request.js
@@ -8,7 +8,7 @@ import type {
 import type {Logger} from "../shared/types.js";
 import {makeRequest} from "./make-request.js";
 import {isFromCache} from "./requests-from-cache.js";
-import {trace} from "../ka-shared/index.js";
+import {trace} from "../shared/index.js";
 
 /**
  * This tracks our inflight requests.
