+++ src/shared/get-request-logger.js
@@ -3,7 +3,14 @@ import type {$Request} from "express";
 import type {Logger, RequestWithLog} from "./types.js";
 
 export const getRequestLogger = <TReq: RequestWithLog<$Request>>(
-    defaultLogger: Logger,
+    defaultLogger: ?Logger,
     request?: TReq,
-): Logger =>
-    request != null && request.log != null ? request.log : defaultLogger;
+): Logger => {
+    if (defaultLogger == null) {
+        if (request == null || request.log == null) {
+            throw new Error("No logs available");
+        }
+        return request.log;
+    }
+    return request != null && request.log != null ? request.log : defaultLogger;
+};
